requirements.txt
pandas
scikit-learn
numpy
mlflow
fastapi
uvicorn
google-cloud-storage
joblib
pyyaml
streamlit
pytest
pytest-cov
python-dotenv
requests
openpyxl
matplotlib
seaborn
httpx
